Padded and non-padded versions (#21)
Introduce the padded version (currently the only one) as a variant, so
the non-padded one can be used for everything else other than the
capricious macOS.


This PR turned out to be trickier than expected, because of how
gradients and coordinates and transforms work interact with each other
in svg. Longs tory short:

- it's a mess, and we have to be careful when making a new logo that it
doesn't use a transform (only quick way to check is to look at the code
itself
- I ended up "solving" the issue with the padded version by ensuring all
logo variants and templates have the same size for the logo itself
(824x824), while also having the top-left corner always at the origin
(which ensures gradients look the same in all variants...). This means
that I had to shift the "viewbox" of the padded logo so that the origin
is not at the top left of the page xD

Bottom line is: when creating new logos, we should always use an
existing one as template and refrain from selecting a group and
moving/rescaling (which might introduce a transform). It's still fixable
if we mess up, but it might give surprising results at first :P