Remove unused defs and clean up templates (#35)
These defs are leftovers from working in inskape, and cause some
warnings when integrating these logos into napari via qt due to
duplicate tags caused by our generation script here. This makes it
easier, and logos are unchanged.

I also cleaned up the templates to their minimal parts which also helps
with this issue, and makes it easier to iterate in the future.